Moved LA file mode for toolbar icon test in test_backends_interactive.py
