+++ lib/matplotlib/tests/test_backends_interactive.py
@@ -7,13 +7,17 @@ import platform
 import signal
 import subprocess
 import sys
+import tempfile
 import time
 import urllib.request
+
+from PIL import Image
 
 import pytest
 
 import matplotlib as mpl
 from matplotlib import _c_internal_utils
+from matplotlib.backend_tools import ToolToggleBase
 from matplotlib.testing import subprocess_run_helper as _run_helper
 
 
@@ -71,6 +75,24 @@ def _get_testable_interactive_backends():
 _test_timeout = 60  # A reasonably safe value for slower architectures.
 
 
+def _test_toolbar_button_la_mode_icon(fig):
+    # test a toolbar button icon using an image in LA mode (GH issue 25174)
+    # create an icon in LA mode
+    with tempfile.TemporaryDirectory() as tempdir:
+        img = Image.new("LA", (26, 26))
+        tmp_img_path = os.path.join(tempdir, "test_la_icon.png")
+        img.save(tmp_img_path)
+
+        class CustomTool(ToolToggleBase):
+            image = tmp_img_path
+            description = ""  # gtk3 backend does not allow None
+
+        toolmanager = fig.canvas.manager.toolmanager
+        toolbar = fig.canvas.manager.toolbar
+        toolmanager.add_tool("test", CustomTool)
+        toolbar.add_tool("test", "group")
+
+
 # The source of this function gets extracted and run in another process, so it
 # must be fully self-contained.
 # Using a timer not only allows testing of timers (on other backends), but is
@@ -122,13 +144,16 @@ def _test_interactive_impl():
         if importlib.util.find_spec("cairocffi"):
             check_alt_backend(backend[:-3] + "cairo")
         check_alt_backend("svg")
-
     mpl.use(backend, force=True)
 
     fig, ax = plt.subplots()
     assert_equal(
         type(fig.canvas).__module__,
         f"matplotlib.backends.backend_{backend}")
+
+    if mpl.rcParams["toolbar"] == "toolmanager":
+        # test toolbar button icon LA mode see GH issue 25174
+        _test_toolbar_button_la_mode_icon(fig)
 
     ax.plot([0, 1], [2, 3])
     if fig.canvas.toolbar:  # i.e toolbar2.
@@ -168,11 +193,17 @@ def test_interactive_backend(env, toolbar):
             pytest.skip("toolmanager is not implemented for macosx.")
     if env["MPLBACKEND"] == "wx":
         pytest.skip("wx backend is deprecated; tests failed on appveyor")
-    proc = _run_helper(_test_interactive_impl,
-                       json.dumps({"toolbar": toolbar}),
-                       timeout=_test_timeout,
-                       extra_env=env)
-
+    try:
+        proc = _run_helper(
+                _test_interactive_impl,
+                json.dumps({"toolbar": toolbar}),
+                timeout=_test_timeout,
+                extra_env=env,
+                )
+    except subprocess.CalledProcessError as err:
+        pytest.fail(
+                "Subprocess failed to test intended behavior\n"
+                + str(err.stderr))
     assert proc.stdout.count("CloseEvent") == 1
